Fix stop_words_for under Ruby 3+
Fixes "no implicit conversion of Hash into Integer" under Ruby 3+

I'm not actually sure how this ever worked properly. Kernel#open apparently
supported the 3rd argument being a hash for options, despite not really being
clear in the docs. Stopped working in Ruby 3.0.

This actual code was for Ruby 1.9 anyway, and we no longer support that so just
ditching it!